requirements.txt
selenium
python-dotenv
pyyaml
behave
